[수정] redis 설정 제거

+++ src/main/kotlin/cc/connectcampus/connect_campus/domain/chat/repository/MessageMongoRepository.kt
@@ -9,5 +9,5 @@ import java.util.*
 
 interface MessageMongoRepository: MongoRepository<Message, String> {
     fun findFirstByChatIdAndSentAtGreaterThanOrderBySentAtDesc(chatId: UUID, joinedAt: LocalDateTime): Message?
-    fun countByChatIdAndSentAtGreaterThanAndUnReadMembersNotContains(chatId: UUID, joinedAt: LocalDateTime, memberId: UUID): Long
+    fun countByChatIdAndSentAtGreaterThanAndUnReadMembersNotContains(chatId: UUID, joinedAt: LocalDateTime, memberId: UUID): Int
 }
